Move monkey patch and behave hooks into separate module
